mainpage cleanup, basic image detail page
