+++ Scanbot.SDK.Example.Forms/Pages/ImageDetailPage.cs
@@ -1,0 +1,10 @@
+﻿using System;
+namespace Scanbot.SDK.Example.Forms
+{
+    public class ImageDetailPage
+    {
+        public ImageDetailPage()
+        {
+        }
+    }
+}
