+++ Scanbot.SDK.Example.Forms/Scanbot.SDK.Example.Forms.projitems
@@ -30,6 +30,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Subviews\Cells\ImageResultCell.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Subviews\ActionBar\BottomActionBar.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Subviews\ActionBar\BottomActionButton.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Pages\ImageDetailPage.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="$(MSBuildThisFileDirectory)Utils\" />
